openIncognito refactor, 'always open in incognito' now works globally
