+++ bbcode/UrlTagView.vue
@@ -20,7 +20,6 @@
 
 <script lang="ts">
   import Vue from 'vue';
-  import core from '../chat/core';
   import { EventBus } from '../chat/preview/event-bus';
 
   export default Vue.extend({
@@ -58,14 +57,8 @@
         if (e.altKey) {
           this.toggleStickyness();
           e.preventDefault();
-          return;
-        }
-
-        this.dismiss(true);
-
-        if (core.state.generalSettings?.horizonAlwaysOpenIncognito) {
-          e.preventDefault();
-          EventBus.$emit('open-url-incognito', { url: this.url });
+        } else {
+          this.dismiss();
         }
       }
     }
